# frontend/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "PaymentSim API",
    "version": "v1"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign in",
        "description": "Verifies email/password and sets the auth cookie. Returns 200 with the user, or 401 if the credentials are wrong.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Register a customer",
        "description": "Creates a Customer account and signs them in. 400 on invalid input, 409 if the email is taken.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign out",
        "description": "Clears the auth cookie. Always returns 204.",
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Current user",
        "description": "Returns the signed-in user, or 401 if there's no active session.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "List all orders",
        "description": "Returns every order, newest first. Use it to see an order flip from Pending to Paid after a webhook.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Start a checkout",
        "description": "Creates a Pending order and a matching Stripe PaymentIntent, then returns the clientSecret the frontend uses to confirm the card.\n\nBody: amountCents (integer, in cents — 1999 = 19.99) and currency (lowercase ISO code, e.g. \"eur\").",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrderResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}/refund": {
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "Refund a paid order",
        "description": "Asks Stripe to refund the order's payment. Returns 202 Accepted immediately; the order flips to Refunded only when the charge.refunded webhook arrives. Fails with 404 if the order doesn't exist, or 400 if it isn't Paid.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefundResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Liveness check",
        "description": "Returns 200 with { status: \"ok\" } if the API is up. No parameters.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/String<>f__AnonymousType5"
                }
              }
            }
          }
        }
      }
    },
    "/webhook": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "Stripe webhook receiver",
        "description": "Called by Stripe (not by the frontend). Requires a valid 'Stripe-Signature' header; the raw body is verified against the webhook secret.\n\nYou cannot produce a valid call from Swagger — use the Stripe CLI (`stripe listen` + `stripe trigger`). Calling it here returns 400 (no valid signature), which demonstrates the forged-request rejection.",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CreateOrderRequest": {
        "type": "object",
        "properties": {
          "amountCents": {
            "type": "integer",
            "description": "Amount in the smallest currency unit (cents). Example: 1999 means 19.99.",
            "format": "int64"
          },
          "currency": {
            "type": "string",
            "description": "Lowercase ISO currency code. Example: \"eur\".",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request body to start a checkout."
      },
      "CreateOrderResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "The new order's id.",
            "format": "uuid"
          },
          "clientSecret": {
            "type": "string",
            "description": "The Stripe clientSecret the frontend uses to confirm the card.",
            "nullable": true
          },
          "amountCents": {
            "type": "integer",
            "description": "Amount in the smallest currency unit (cents).",
            "format": "int64"
          },
          "currency": {
            "type": "string",
            "description": "Lowercase ISO currency code, e.g. \"eur\".",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Response returned when a checkout starts."
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Account email (case-insensitive).",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "Account password (plaintext over HTTPS; verified against a stored hash).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Credentials for POST /auth/login."
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "amountCents": {
            "type": "integer",
            "format": "int64"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "stripePaymentIntentId": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "OrderStatus": {
        "enum": [
          "Pending",
          "Paid",
          "Failed",
          "Refunded"
        ],
        "type": "string"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "RefundResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "The order being refunded.",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "description": "Always \"refund_pending\" — the order flips to Refunded only on the charge.refunded webhook.",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "Human-readable explanation the frontend can show.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Response returned when a refund is accepted (still pending)."
      },
      "RegisterRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The new account's email (must be unique).",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "The new account's password (min 6 chars).",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Details for POST /auth/register."
      },
      "String<>f__AnonymousType5": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "The user's email.",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          }
        },
        "additionalProperties": false,
        "description": "The signed-in user, returned by login and /auth/me."
      },
      "UserRole": {
        "enum": [
          "Admin",
          "Customer"
        ],
        "type": "string"
      }
    }
  },
  "tags": [
    {
      "name": "Auth"
    },
    {
      "name": "Orders"
    },
    {
      "name": "System"
    },
    {
      "name": "Webhooks"
    }
  ]
}